- Converted project to Visual Studio 2013 - Providing 64 bit configuration

+++ JPEGView/EXIFReader.cpp
@@ -54,7 +54,7 @@ static void ReadStringTag(CString & strOut, uint8* ptr, uint8* pTIFFHeader, bool
 		if (ptr != NULL && ReadUShort(ptr + 2, bLittleEndian) == 2) {
 			int nSize = ReadUInt(ptr + 4, bLittleEndian);
 			LPCSTR pString = (nSize <= 4) ? (LPCSTR)(ptr + 8) : (LPCSTR)(pTIFFHeader + ReadUInt(ptr + 8, bLittleEndian));
-            int nMaxChars = min(nSize, strlen(pString));
+			int nMaxChars = min(nSize, (int)strlen(pString));
             if (bTryReadAsUTF8) {
                 CString strUTF8 = Helpers::TryConvertFromUTF8((uint8*)pString, nMaxChars);
                 if (!strUTF8.IsEmpty()) {
